VIDSOL-300 screen sharing module - cleanup and minor changes from review
