+++ vonage-feature-screensharing/src/enabled/java/com/vonage/android/screensharing/di/ScreenSharingModule.kt
@@ -12,7 +12,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object ScreensharingModule {
+object ScreenSharingModule {
 
     @Provides
     @Singleton
